restore config files as data classes bump infinitic version to 0.12.2 rename InfiniticConsumer to InfiniticConsumerAsync
