+++ infinitic-client/src/main/kotlin/io/infinitic/clients/dispatcher/ClientDispatcher.kt
@@ -52,7 +52,7 @@ import io.infinitic.common.proxies.RequestByWorkflowTag
 import io.infinitic.common.tasks.data.ServiceName
 import io.infinitic.common.tasks.data.TaskId
 import io.infinitic.common.tasks.executors.errors.MethodFailedError
-import io.infinitic.common.transport.InfiniticConsumer
+import io.infinitic.common.transport.InfiniticConsumerAsync
 import io.infinitic.common.transport.InfiniticProducer
 import io.infinitic.common.workflows.data.channels.SignalId
 import io.infinitic.common.workflows.data.methodRuns.MethodRunId
@@ -101,7 +101,7 @@ import io.infinitic.common.workflows.tags.messages.RetryTasksByTag as RetryTaskI
 
 class ClientDispatcher(
   logName: String,
-  private val consumer: InfiniticConsumer,
+  private val consumerAsync: InfiniticConsumerAsync,
   private val producer: InfiniticProducer
 ) : ProxyDispatcher, Closeable {
   private val logger = KotlinLogging.logger(logName)
@@ -741,7 +741,7 @@ class ClientDispatcher(
     // lazily starts client consumer if not already started
     synchronized(this) {
       if (!isClientConsumerInitialized) {
-        consumer.startClientConsumerAsync(::handle, null, clientName)
+        consumerAsync.startClientConsumerAsync(::handle, null, clientName)
         isClientConsumerInitialized = true
       }
     }
