+++ infinitic-client/src/test/kotlin/io/infinitic/clients/ClientWorkflowTests.kt
@@ -43,7 +43,7 @@ import io.infinitic.common.fixtures.TestFactory
 import io.infinitic.common.fixtures.later
 import io.infinitic.common.tasks.executors.messages.TaskExecutorMessage
 import io.infinitic.common.tasks.tags.messages.TaskTagMessage
-import io.infinitic.common.transport.InfiniticConsumer
+import io.infinitic.common.transport.InfiniticConsumerAsync
 import io.infinitic.common.transport.InfiniticProducerAsync
 import io.infinitic.common.workflows.data.channels.ChannelName
 import io.infinitic.common.workflows.data.channels.ChannelType
@@ -128,11 +128,11 @@ private val producerAsync: InfiniticProducerAsync = mockk<InfiniticProducerAsync
   every { sendAsync(capture(workflowEngineSlot), capture(delaySlot)) } answers { engineResponse() }
 }
 
-private val consumer: InfiniticConsumer = mockk<InfiniticConsumer> {
+private val consumerAsync: InfiniticConsumerAsync = mockk<InfiniticConsumerAsync> {
   every { startClientConsumerAsync(any(), any(), clientNameTest) } returns completed()
 }
 
-private val client = InfiniticClient(consumer, producerAsync)
+private val client = InfiniticClient(consumerAsync, producerAsync)
 
 private class ClientWorkflowTests : StringSpec(
     {
@@ -175,7 +175,7 @@ private class ClientWorkflowTests : StringSpec(
         )
 
         // when asynchronously dispatching a workflow, the consumer should not be started
-        verify { consumer.startClientConsumerAsync(any(), any(), any()) wasNot called }
+        verify { consumerAsync.startClientConsumerAsync(any(), any(), any()) wasNot called }
       }
 
       "Should be able to dispatch a workflow with annotation" {
@@ -379,13 +379,13 @@ private class ClientWorkflowTests : StringSpec(
         )
 
         // when waiting for a workflow, the consumer should be started
-        verify { consumer.startClientConsumerAsync(any(), any(), clientNameTest) }
+        verify { consumerAsync.startClientConsumerAsync(any(), any(), clientNameTest) }
 
         // restart a workflow
         client.dispatch(fakeWorkflow::m3, 0, "a").await()
 
         // the consumer should be started only once
-        verify { consumer.startClientConsumerAsync(any(), any(), any()) wasNot called }
+        verify { consumerAsync.startClientConsumerAsync(any(), any(), any()) wasNot called }
       }
 
       "Should throw a WorkflowTimedOutException when waiting for a workflow more than timeout" {
