+++ infinitic-common/src/main/resources/schemas/taskTagEnvelope-0.12.2.avsc
@@ -1,0 +1,117 @@
+{
+  "type" : "record",
+  "name" : "TaskTagEnvelope",
+  "namespace" : "io.infinitic.tasks.tag",
+  "fields" : [ {
+    "name" : "name",
+    "type" : "string"
+  }, {
+    "name" : "type",
+    "type" : {
+      "type" : "enum",
+      "name" : "TaskTagMessageType",
+      "symbols" : [ "ADD_TAG_TO_TASK", "REMOVE_TAG_FROM_TASK", "CANCEL_TASK_BY_TAG", "RETRY_TASK_BY_TAG", "GET_TASK_IDS_BY_TAG" ]
+    }
+  }, {
+    "name" : "addTagToTask",
+    "type" : [ "null", {
+      "type" : "record",
+      "name" : "AddTagToTask",
+      "fields" : [ {
+        "name" : "messageId",
+        "type" : "string"
+      }, {
+        "name" : "taskName",
+        "type" : "string"
+      }, {
+        "name" : "taskTag",
+        "type" : "string"
+      }, {
+        "name" : "taskId",
+        "type" : "string"
+      }, {
+        "name" : "emitterName",
+        "type" : "string"
+      } ]
+    } ]
+  }, {
+    "name" : "removeTagFromTask",
+    "type" : [ "null", {
+      "type" : "record",
+      "name" : "RemoveTagFromTask",
+      "fields" : [ {
+        "name" : "messageId",
+        "type" : "string"
+      }, {
+        "name" : "taskName",
+        "type" : "string"
+      }, {
+        "name" : "taskTag",
+        "type" : "string"
+      }, {
+        "name" : "taskId",
+        "type" : "string"
+      }, {
+        "name" : "emitterName",
+        "type" : "string"
+      } ]
+    } ]
+  }, {
+    "name" : "cancelTaskByTag",
+    "type" : [ "null", {
+      "type" : "record",
+      "name" : "CancelTaskByTag",
+      "fields" : [ {
+        "name" : "messageId",
+        "type" : "string"
+      }, {
+        "name" : "taskName",
+        "type" : "string"
+      }, {
+        "name" : "taskTag",
+        "type" : "string"
+      }, {
+        "name" : "emitterName",
+        "type" : "string"
+      } ]
+    } ]
+  }, {
+    "name" : "retryTaskByTag",
+    "type" : [ "null", {
+      "type" : "record",
+      "name" : "RetryTaskByTag",
+      "fields" : [ {
+        "name" : "messageId",
+        "type" : "string"
+      }, {
+        "name" : "taskName",
+        "type" : "string"
+      }, {
+        "name" : "taskTag",
+        "type" : "string"
+      }, {
+        "name" : "emitterName",
+        "type" : "string"
+      } ]
+    } ]
+  }, {
+    "name" : "getTaskIdsByTag",
+    "type" : [ "null", {
+      "type" : "record",
+      "name" : "GetTaskIdsByTag",
+      "fields" : [ {
+        "name" : "messageId",
+        "type" : "string"
+      }, {
+        "name" : "taskName",
+        "type" : "string"
+      }, {
+        "name" : "taskTag",
+        "type" : "string"
+      }, {
+        "name" : "emitterName",
+        "type" : "string"
+      } ]
+    } ]
+  } ]
+}
