+++ infinitic-common/src/main/resources/versions
@@ -23,3 +23,4 @@
 0.11.7
 0.12.0
 0.12.1
+0.12.2
