+++ infinitic-worker/build.gradle.kts
@@ -30,10 +30,10 @@ dependencies {
   api(project(":infinitic-storage"))
   api(project(":infinitic-cache"))
   api(project(":infinitic-transport"))
+  api(project(":infinitic-client"))
 
   implementation(project(":infinitic-autoclose"))
   implementation(project(":infinitic-common"))
-  implementation(project(":infinitic-client"))
   implementation(project(":infinitic-transport"))
   implementation(project(":infinitic-task-executor"))
   implementation(project(":infinitic-workflow-engine"))
