+++ infinitic-worker/src/main/kotlin/io/infinitic/workers/config/WorkerConfigInterface.kt
@@ -20,9 +20,15 @@
  *
  * Licensor: infinitic.io
  */
-package io.infinitic.cache.config
+package io.infinitic.workers.config
 
-/** Cache configuration */
-interface CacheConfig {
-  val cache: Cache
-}
+import io.infinitic.cache.config.CacheConfigInterface
+import io.infinitic.clients.config.ClientConfigInterface
+import io.infinitic.storage.config.StorageConfigInterface
+import io.infinitic.workers.register.config.RegisterConfigInterface
+
+interface WorkerConfigInterface :
+  RegisterConfigInterface,
+  CacheConfigInterface,
+  StorageConfigInterface,
+  ClientConfigInterface
